22-12-2006. Commision of latest sfxc version after extensive tests on jop32 (autos) and huygens (crosses). Everything seems ok. The only real test left is comparing the effect of the delay with a reference
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@60 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/staPrms.h
@@ -58,7 +58,7 @@ class StaP
     //delay correction related data
     char *delaytable; //table with delay data
     char *phasetable; //table with phase
-    int  loobs;       //local ascilator observing station
+    INT64 loobs;       //local ascilator observing station
 
     //private functions
     int findMK4data(FILE *ctrlP);
@@ -96,7 +96,7 @@ class StaP
     char* get_modpat();
     char* get_delaytable();
     char* get_phasetable();
-    int   get_loobs();
+    INT64 get_loobs();
     int*  get_signBS(); //bit shift parameters for sign
     int*  get_magnBS(); //bit shift parameters for magn
